Restructured Classes
Restructured classes to allow for easy additions
