+++ src/HangmanDriver.java
@@ -1,4 +1,5 @@
 public class HangmanDriver {
+    
     public static void main(String[] args) {
         // Create an instance of Main Menu
         OpeningMenu hangmanMenu = new OpeningMenu();
