+++ src/HangmanUI.java
@@ -1,3 +1,9 @@
+import java.awt.BorderLayout;
+import java.awt.FlowLayout;
+import java.awt.Font;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
 import javax.swing.ImageIcon;
 import javax.swing.JButton;
 import javax.swing.JFrame;
@@ -5,19 +11,39 @@ import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
 
-import java.awt.BorderLayout;
-import java.awt.FlowLayout;
-import java.awt.Font;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
 public class HangmanUI {
-    public static void main(String[] args) {
+
+    private HangmanModel hm;
+    private int currentLevel;
+
+    public HangmanUI(int currentLevel) {
+        hm = new HangmanModel();
+        hm.hangmanRound();
+        this.currentLevel = currentLevel;
+
+    }
+    public void initalizeUI() {
         //Frame that holds everything
         JFrame frame = new JFrame("Hangman GUI");
 
-        //Panel for the level, navigation controls, pause/play
+        //Set layout manager for the frame
+        frame.setLayout(new BorderLayout());
+
+        // Add the panels to the frame
+        frame.add(createTopPanel(), BorderLayout.NORTH);
+        frame.add(createCenterPanel(), BorderLayout.CENTER);
+        frame.add(createBottomPanel(), BorderLayout.SOUTH);
+
+        // Set frame properties
+        frame.setSize(500, 700); // Set size
+        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE); // Close operation
+        frame.setLocationRelativeTo(null); // Center the frame
+        frame.setVisible(true); // Make the frame visible
+    }
+
+    private JPanel createTopPanel() {
         JPanel topPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 80, 10));
+
             ImageIcon backIcon = new ImageIcon("./images/angle-left.png");
             JButton back = new JButton(backIcon);
             back.addActionListener(new ActionListener() {
@@ -27,9 +53,10 @@ public class HangmanUI {
                     // In the past, I have used a layered pane to swap between them
                 }
             }); 
+
             topPanel.add(back);
 
-            JLabel level = new JLabel("Level 1");
+            JLabel level = new JLabel("Level " + Integer.toString(currentLevel));
             level.setFont(new Font("Sans-serif", Font.BOLD, 14));
             topPanel.add(level);
 
@@ -42,30 +69,47 @@ public class HangmanUI {
             });
             topPanel.add(pause);
 
+            return topPanel;
+    }
+
+    private JPanel createCenterPanel() {
         //Panel for the stick figure and word
         JPanel centerPanel = new JPanel();
-            ImageIcon stickFigureIcon = new ImageIcon("./images/example.png");
-            JLabel exampleStickFigure = new JLabel(stickFigureIcon);
-            centerPanel.add(exampleStickFigure);
-            /* Once we integrate all of the current branches to main, using the random word getter
-            class to generate our word whenever you press the play button and then we get the correct
-            number of blanks back to populate this JLabel */
-            JLabel wordBlanks = new JLabel("_ _ _ _ _ _");
-            wordBlanks.setFont(new Font("Sans-serif", Font.BOLD, 60));
-            centerPanel.add(wordBlanks);
+            centerPanel.add(createStickFigure());
+            centerPanel.add(createWordDisplay());
+            centerPanel.add(createAlphabetPanel());
+            
+            return centerPanel;
 
-        //Panel for the alphabet 
+    }
+
+    private JLabel createStickFigure() {
+        ImageIcon stickFigureIcon = new ImageIcon("./images/example.png");
+        JLabel exampleStickFigure = new JLabel(stickFigureIcon);
+        
+        return exampleStickFigure;
+    }
+
+    private JLabel createWordDisplay() {
+        JLabel wordBlanks = new JLabel("_ _ _ _ _ _");
+        wordBlanks.setFont(new Font("Sans-serif", Font.BOLD, 60));
+        
+        return wordBlanks;
+    }
+
+    private JPanel createAlphabetPanel() {
         JPanel alphabetPanel = new JPanel(new BorderLayout(20, 0));
-            JLabel alphabetRow1 = new JLabel("a b c d e f g h i j k l m");
-            alphabetRow1.setFont(new Font("Sans-serif", Font.BOLD, 30));
-            alphabetPanel.add(alphabetRow1, BorderLayout.NORTH);
-            JLabel alphabetRow2 = new JLabel("n o p q r s t u v w x y z");
-            alphabetRow2.setFont(new Font("Sans-serif", Font.BOLD, 30));
-            alphabetPanel.add(alphabetRow2, BorderLayout.SOUTH);
+        JLabel alphabetRow1 = new JLabel("a b c d e f g h i j k l m");
+        alphabetRow1.setFont(new Font("Sans-serif", Font.BOLD, 30));
+        alphabetPanel.add(alphabetRow1, BorderLayout.NORTH);
+        JLabel alphabetRow2 = new JLabel("n o p q r s t u v w x y z");
+        alphabetRow2.setFont(new Font("Sans-serif", Font.BOLD, 30));
+        alphabetPanel.add(alphabetRow2, BorderLayout.SOUTH);
+        
+        return alphabetPanel;
+    }
 
-        //Put alphabetPanel inside centerPanel
-        centerPanel.add(alphabetPanel);
-
+    private JPanel createBottomPanel() {
         //Panel for letter input
         JPanel bottomPanel = new JPanel();
             JLabel guessLabel= new JLabel("Guess a letter!");
@@ -86,18 +130,6 @@ public class HangmanUI {
             });
             bottomPanel.add(guessField);
 
-        //Set layout manager for the frame
-        frame.setLayout(new BorderLayout());
-
-        // Add the panels to the frame
-        frame.add(topPanel, BorderLayout.NORTH);
-        frame.add(centerPanel, BorderLayout.CENTER);
-        frame.add(bottomPanel, BorderLayout.SOUTH);
-
-        // Set frame properties
-        frame.setSize(500, 700); // Set size
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE); // Close operation
-        frame.setLocationRelativeTo(null); // Center the frame
-        frame.setVisible(true); // Make the frame visible
+            return bottomPanel;
     }
 }
