+++ tests/HangmanUITester.java
@@ -1,0 +1,21 @@
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+
+import org.junit.Test;
+
+
+
+public class HangmanUITester {
+    private ByteArrayInputStream inputStream;
+    private InputStream originalSystemIn;
+
+    @Test
+    public void testWin() {
+        
+    }
+
+    public static void main(String[] args) {
+        HangmanUI hu = new HangmanUI(1);
+    }
+
+}
