+++ tests/ModelTester.java
@@ -1,24 +1,23 @@
 
 
 
-import org.junit.Test;
-
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 
+import org.junit.Test;
 
 
-public class ControllerTester {
+
+public class ModelTester {
     private ByteArrayInputStream inputStream;
     private InputStream originalSystemIn;
 
     @Test
     public void testWin() {
         textInputs("a\nm\nz\ni\nn\ng\n"); //amazing is default word
-        HangmanController hc = new HangmanController();
+        HangmanModel hc = new HangmanModel();
         hc.hangmanRound();
         assertTrue(hc.isWordGuessed());
         System.setIn(originalSystemIn);
@@ -28,7 +27,7 @@ public class ControllerTester {
     @Test
     public void testFail() {
         textInputs("b\nc\nd\ne\nf\nh\nj\n"); //7 tries 
-        HangmanController hc = new HangmanController();
+        HangmanModel hc = new HangmanModel();
         hc.hangmanRound();
         assertFalse(hc.isWordGuessed());
         System.setIn(originalSystemIn);
@@ -42,7 +41,7 @@ public class ControllerTester {
     }
 
     public static void main(String[] args) {
-        HangmanController hc = new HangmanController();
+        HangmanModel hc = new HangmanModel();
         hc.hangmanRound(); 
     }
